requirements.txt
# Core Framework
streamlit==1.28.1
langchain==0.0.340
langchain-google-genai==0.0.7

# Data Processing
pandas==2.1.3
numpy==1.26.4
openpyxl==3.1.2
python-docx==0.8.11

# LLM Integration
google-generativeai==0.3.2


# Database
sqlalchemy==2.0.23

# Visualization
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Data Quality
great-expectations==0.18.8

# Utilities
python-dotenv==1.0.0
pydantic==1.10.13
typing-extensions==4.8.0
zipfile36==0.1.3

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# PDF generation
fpdf==1.7.2


